+++ modules/hivemq-edge-module-opcua/src/main/java/com/hivemq/edge/adapters/opcua/OpcUaProtocolAdapter.java
@@ -258,7 +258,7 @@ public class OpcUaProtocolAdapter implements WritingProtocolAdapter {
             return;
         }
 
-        final int healthCheckIntervalSeconds = config.getConnectionOptions().healthCheckInterval();
+        final long healthCheckIntervalSeconds = config.getConnectionOptions().healthCheckInterval();
         final ScheduledFuture<?> future = healthCheckScheduler.scheduleAtFixedRate(() -> {
             // Check if adapter was stopped before health check executes
             if (stopped) {
@@ -507,7 +507,7 @@ public class OpcUaProtocolAdapter implements WritingProtocolAdapter {
                 this.opcUaClientConnection.set(null);
                 protocolAdapterState.setConnectionStatus(ProtocolAdapterState.ConnectionStatus.ERROR);
 
-                final int retryIntervalSeconds = config.getConnectionOptions().retryInterval();
+                final long retryIntervalSeconds = config.getConnectionOptions().retryInterval();
                 if (throwable != null) {
                     log.warn("OPC UA adapter '{}' connection failed, will retry in {} seconds",
                             adapterId, retryIntervalSeconds, throwable);
@@ -535,7 +535,7 @@ public class OpcUaProtocolAdapter implements WritingProtocolAdapter {
             return;
         }
 
-        final int retryIntervalSeconds = config.getConnectionOptions().retryInterval();
+        final long retryIntervalSeconds = config.getConnectionOptions().retryInterval();
         final ScheduledFuture<?> future = retryScheduler.schedule(() -> {
             // Check if adapter was stopped before retry executes
             if (stopped || this.parsedConfig == null || this.moduleServices == null) {
